+++ src/battle/view/EnemyView.java
@@ -8,7 +8,7 @@ package battle.view;
  */
 public class EnemyView extends ItemView {
 	private static final String ENEMY_NAME = "Enemy";
-	public EnemyView(int hp, int x, int y, String filePath) {
+	public EnemyView(double hp, int x, int y, String filePath) {
 		super(ENEMY_NAME, hp, x, y,filePath);
 	}
 }
