+++ src/battle/view/PlayerView.java
@@ -9,7 +9,7 @@ package battle.view;
 public class PlayerView extends ItemView {
 	private static final String PLAYER_NAME = "Player";
 
-	public PlayerView(int hp, int x, int y,String filePath) {
+	public PlayerView(double hp, int x, int y,String filePath) {
 		super(PLAYER_NAME, hp, x, y,filePath);
 	}
 }
